Move Create General Email handlers into their own package
